test: 변경된 것에 맞게 인수 테스트 수정

+++ src/test/java/nextstep/subway/documentation/PathDocumentation.java
@@ -31,7 +31,7 @@ public class PathDocumentation extends Documentation {
                 ),
                 10,
                 10,
-                BigDecimal.valueOf(1250)
+                1250
         );
 
         when(pathService.findPath(1L, 2L, PathType.DISTANCE)).thenReturn(pathResponse);
